add MovieDetailsGridContent
